Object attributes dialog fill and stroke selector
This change adds popup for setting object fill and stroke paint.
Object attributes supplements this with stroke attributes UI for
adjusting stroke thickness, paint order, markers, miter limit, etc.

* Object fill stroke attributes popup edit
* HSL color wheel modifications
* HSL color wheel improvements based on Henrique's proposal "Inkscape UI Vision"
* Add property grid to object attributes dialog
* Gradient editor improvements and fixes
* Implemented simplified gradient editor layout as proposed in
https://gitlab.com/inkscape/ux/-/issues/246
* Fixed gradient editor handles positioning and rounding.
* Implementing UX 246 style popup
* Color selector popup for flat colors and grid editor improvements to match UX 246 design.
* Set object attributes from paint selector popup
* Object attributes to use new paint selector
* Object attribute dialog improvements
* Object attributes - editing stroke caps/joins/order
* Implemented stoke options popup with relevant stroke attributes details
* Pattern editing fixes
* SignalObserver extended to provide obj change details
* Swatch editor for popup paint selector
* Add OkHsv color using B. Ottosson's implementation (it's fast)
* Swatch editor
* Support multiobject selection in properties dialog (preliminary)
* Add basic support for multiobject selection.
* Scan selected objects, collect element types and fill paints (test)
* Added missing icons
* Restored color wheel
* Added preference saving in paint switch
* Add out of sRGB gamut warning
* Restore gradient editor color wheel
* Try to reuse swatches instead of creating new ones
* Improve filter reporting and editing
* Fix color gamut and okhsv conversion
* Update marker popup, add opacity
* Fix changing paint order
* Fix rbg test
* Stop rounttrip when assigning flat color
* Add to-css-gamut functionality
* Fix xyz and lch color tests
* Add tags to sp-object modified flags
* Fix double update in object properties anchor element
* Fix UI refresh when clearing paint
* Enhance image panel in object properties
* Reset blend and opacity improvements
* Fix color comparison precision
* Add missing tooltips
* Add dash selector and marker combo tooltips
* Allow spinbtn to shrink
* Marker popup redesign
* Pattern popup redesign

+++ src/style-internal.cpp
@@ -2300,8 +2300,8 @@ bool SPIDashArray::is_valid() const {
 // SPIFontSize ----------------------------------------------------------
 
 /** Indexed by SP_CSS_FONT_SIZE_blah.   These seem a bit small */
-float const SPIFontSize::font_size_table[] = {6.0, 8.0, 10.0, 12.0, 14.0, 18.0, 24.0};
-float const SPIFontSize::font_size_default = 12.0;
+double const SPIFontSize::font_size_table[] = {6.0, 8.0, 10.0, 12.0, 14.0, 18.0, 24.0};
+double const SPIFontSize::font_size_default = 12.0;
 
 void
 SPIFontSize::read( gchar const *str ) {
